requirements.txt
Flask==3.0.3
Jinja2==3.1.6
Werkzeug==3.0.6
blinker==1.8.2
markupsafe==2.1.5
itsdangerous==2.2.0
click==8.1.8
importlib-metadata==8.5.0
zipp==3.20.2
pandas
requests
folium
jieba
volcenginesdkarkruntime
python-dotenv 